update approval notifier message with PR status
This commit adds a new PR status section to the approval notifier message.
This section shows the status of approvals of the folders involved in the PR.
It shows if all the files under a folder are completely approved, partially approved
or unapproved. The folder used are the folder where the owners files of the assocaited
with the original Pr reside.